services/face-service-2D-simple/pyproject.toml
[project]
name = "face-service-2D-simple"
version = "0.1.0"
description = "2D VTuber face overlay with WebSocket integration"
requires-python = ">=3.10"
dependencies = ["PySide6>=6.7", "pydantic>=2.7", "websockets>=12.0", "PyYAML>=6.0"]
